版本1.0.0.6 使用JPEG流图片作为车尾图片

+++ VehRecDll/VehRecDll/CameraModule/CameraResult.h
@@ -106,8 +106,8 @@ public:
     float fConfidenceLevel;     //©ипе╤х
 
     UINT64 dw64TimeMS;
-    DWORD dwReceiveTime;
-	UINT32 dwCarID;
+	ULONG dwReceiveTime;
+	ULONG dwCarID;
     INT32 iDeviceID;
     INT32 iPlateColor;
     INT32 iPlateTypeNo;
